Implement new system for styling hyperlinks
While I originally believed that I could eliminate most custom properties for hyperlink styling, that didn't work out in practice. That was discarded prototype #1 today.

Based on https://www.sitepoint.com/css-theming-custom-properties-javascript/ and https://www.smashingmagazine.com/2018/05/css-custom-properties-strategy-guide/, I decided to switch from global color properties to localized ones. However, a prototype relying on only three variables—for foreground, underline, and background—did NOT work at all. That became discarded prototype #2 today.

The reason that three properties are insufficient is that :focus and :hover styles need to be able to partially override :link and :visited styles. Furthermore, :active styles need to do the same for :focus and :hover, thus forming a very special cascade.

Unfortunately, the working solution requires nine properties, three for each hyperlink pseudostate. The reason for this not becoming onerous in practice is how custom properties are read. For each pseudostate, the var() calls reading a custom property for its color value also has a fallback. For the :focus and :hover pseudo-state, the fallback simply is the right value for the default color theme. But for the :active pseudostate, the default is another var() call, which tries to access the previous state's corresponding property. Only if that isn't defined either does the nested call fall back on the value for the default theme. Do You Be, Ubu Trump, and So Long all have different hyperlink color themes. So Long has two even. Yet they only require 2 or 5 property definitions. As an added benefit, they can be made on an ancestor element and do not require the developer to rewrite the pseudostate selectors.


Former-commit-id: e26946ed365386ed0c61073dded7429610f4eafe