Update privacy policy with more costfree restrictions and better typography
Google is surfacing more and more options on the Analytics website that makes me responsible for Google's privacy-destroying ways. Notably, I didn't share data with Google and have no plans to do so in the future—except I want that option for diagnosing technical issues but not for on-going monitoring. So I spelled the non-sharing out in the policy and left the door open for using it in the future temporarily.


Former-commit-id: e23718d5b0639cf17099f05c4ddba4d9494545dd