Capture snapshot of .carddeck and .card
Yikes, this has been much harder than anticipated because:

(1) Using grid auto-layout for a left-to-right language yet wanting cards to be right-aligned is not possible. I am trying to build a much simpler flex style but that will require two more iterations.

(2) Safari does not accept anchors that do *not* contain immediate children with text as actual links—even if the href is present and valid. In short,
    <h3><a><span>....</span></a></h3>
does not work in Safari. Adding a real space, as in
    <h3><a>&nbsp;.<span>....</span></a></h3>
solves the problem but also results in layout problems due to the space. The solution, in turn, is to use <a> for making the card a link via absolute positionening absolute with zero margins all around. The nested <span> can
still serve as visible text and is absolutely positioned as well, albeit in a nested context.

(3) When absolutely positioning an <img> on the enclosing aspect ratio frame, we cannot use zero margins all around. If we do, then object-fit and object-positioin do not work. For those to be effective, we need to use an explicit width and height.


Former-commit-id: ea5c29146e924099917ced5873ed2e5169f54252