Behold, the first alpha version of Apparebit v3
Some of the changes snuck into the commit two commits back in time. Like this commit, they mostly are comments and clean-up. But the prior one also included the critical bug fix to finally make cards work. As it turns out,. Safari and Chrome just pretend to accept "end" for justify-contents. They most certainly don't act on it and so I am sticking with "flex-end" for now.


Former-commit-id: 9c056ffebbbb65d3619a9ef714efa264fb9654f4