Ensure gradient background for pullquotes works on Chrome
The original styling used "background-attachment: fixed," which still is the more general soolution. On Chrome, however, it does not render unless it is in the visible viewport and the browser window is reloaded or resized, thus triggering a layout. This version works on Chrome as well as Safari and Firefox. It exploits that the gradient is horizontal and sizes the gradient to appear uniform.